+++ R/cff-class.R
@@ -1,0 +1,12 @@
+#' The `cff` class
+#'
+#' @name cff-class
+#'
+#' @description
+#' TODO
+#'
+#' @keywords internal
+#'
+#'
+#'
+NULL