+++ R/cff_parse_person.R
@@ -60,6 +60,10 @@
 cff_parse_person <- function(person) {
   person <- as.person(person)
 
+  if (length(person) == 0) {
+    return(NULL)
+  }
+
   if (length(person) > 1) {
     person <- lapply(person, cff_parse_person)
     class(person) <- "cff"
@@ -68,7 +72,7 @@ cff_parse_person <- function(person) {
 
   # Special case for Bioconductor
 
-  if (is.substring(person$given, "Bioconductor")) {
+  if (is_substring(person$given, "Bioconductor")) {
     person <- person(
       given = paste(
         clean_str(person$given),
@@ -81,11 +85,9 @@ cff_parse_person <- function(person) {
   }
 
   # Special case for R Core Team
-
-
   if (all(
-    is.substring(clean_str(person$given), "R Core"),
-    is.substring(person$family, "Team")
+    is_substring(clean_str(person$given), "R Core"),
+    is_substring(person$family, "Team")
   )) {
     person <- person(
       given = paste(
@@ -112,7 +114,7 @@ cff_parse_person <- function(person) {
   }
 
   # Check if several mails (MomTrunc 6.0)
-  valid_emails <- unlist(lapply(person$email, is.email))
+  valid_emails <- unlist(lapply(person$email, is_email))
   email <- person$email[valid_emails][1]
   parsed_person$email <- clean_str(email)
 
@@ -145,7 +147,7 @@ cff_parse_person <- function(person) {
   web <- parsed_comments$website
 
   if (!is.null(web)) {
-    parsed_comments$website <- clean_str(web[is.url(web)])
+    parsed_comments$website <- clean_str(web[is_url(web)])
   }
 
   # Add comments