+++ tests/testthat/test-utils.R
@@ -24,3 +24,20 @@ test_that("Try cleaning string", {
   expect_null(clean_str(logical(0)))
   expect_null(clean_str(list()))
 })
+
+test_that("Use right repo", {
+  # Use some other repos
+  repos <- c("https://ropensci.r-universe.dev", "https://cloud.r-project.org")
+
+  final <- detect_repos(repos)
+  expect_identical(final, c(CRAN = "https://cloud.r-project.org/"))
+
+  # With posit package manager
+  repos <- c(
+    CRAN = "https://packagemanager.posit.co/cran/latest",
+    RPSM = "https://packagemanager.rstudio.com/",
+    ANOTHER = "https://packagemanager.rspm.com/"
+  )
+  final <- detect_repos(repos)
+  expect_identical(final, c(CRAN = "https://cloud.r-project.org/"))
+})